[FEAT] 문의 페이지

+++ gearshift/src/main/java/com/fastcampus/gearshift/controller/admin/NAdminUserController.java
@@ -1,5 +1,6 @@
 package com.fastcampus.gearshift.controller.admin;
 
+import com.fastcampus.gearshift.dto.AdminDto;
 import com.fastcampus.gearshift.dto.UserDto;
 import com.fastcampus.gearshift.service.NUserListService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +9,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @Controller
@@ -19,7 +21,14 @@ public class NAdminUserController {
 
     // 회원 목록 조회
     @GetMapping("")
-    public String getList(Model model) {
+    public String getList(Model model, HttpSession session) {
+        AdminDto admin = (AdminDto) session.getAttribute("adminUser");
+
+        if (admin == null) {
+            session.setAttribute("redirectAfterLogin", "/admin/userList");
+            return "redirect:/login.do";
+        }
+
         List<UserDto> list = userListService.userList();
         model.addAttribute("userList", list);
         return "admin/userList";
@@ -28,7 +37,15 @@ public class NAdminUserController {
     // 관리자에 의한 회원 삭제
     @PostMapping("/delete")
     public String deleteUser(@RequestParam("userId") Integer userId,
+                             HttpSession session,
                              RedirectAttributes redirectAttributes) {
+
+        AdminDto admin = (AdminDto) session.getAttribute("adminUser");
+        if (admin == null) {
+            session.setAttribute("redirectAfterLogin", "/admin/userList");
+            return "redirect:/login.do";
+        }
+
         try {
             userListService.deleteUser(userId);
             redirectAttributes.addFlashAttribute("message", "회원 삭제가 완료되었습니다.");
